requirements.txt
ollama
requests
playwright
python-dotenv
flask